Update README.md
This is a desktop app with a simple gui built using python libraries. This is basically used provide information about the stock prices of a particular company which the user searches for. It collects information from internet and calculates the change percentage in the stock prices over the last 10 days and sends an appropriate message via an sms to the user. All you have to do is to provide the name of the company the rest is taken care by the bot. It also provides a brief news about the company in the sms. This app can send an sms only to my number as i am using a free trial version of the messenger api. I think this will be of great use for those who invest in the stocks.